Split extension methods into multiple files
