+++ Source/Testably.Architecture.Rules/Filters/FieldFilterExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq.Expressions;
 using System.Reflection;
 
 namespace Testably.Architecture.Rules;
@@ -6,53 +7,29 @@ namespace Testably.Architecture.Rules;
 /// <summary>
 ///     Extension fields for <see cref="IFieldFilter" />.
 /// </summary>
-public static class FieldFilterExtensions
+public static partial class FieldFilterExtensions
 {
 	/// <summary>
-	///     Filter <see cref="FieldInfo" />s that have an attribute of type <typeparamref name="TAttribute" />.
+	///     Filters the applicable <see cref="FieldInfo" /> on which the expectations should be applied.
 	/// </summary>
 	/// <param name="this">The <see cref="IFieldFilter" />.</param>
-	/// <param name="predicate">
-	///     (optional) A predicate to check the attribute values.
-	///     <para />
-	///     If not set (<see langword="null" />), will only check if the attribute is present.
-	/// </param>
-	public static WithAttributeFilterResult WithAttribute<TAttribute>(
-		this IFieldFilter @this,
-		Func<TAttribute, bool>? predicate = null)
-		where TAttribute : Attribute
+	/// <param name="filter">The filter to apply on the <see cref="FieldInfo" />.</param>
+	public static IFieldFilterResult Which(this IFieldFilter @this,
+		Expression<Func<FieldInfo, bool>> filter)
 	{
-		WithAttributeFilterResult filter = new(
-			@this);
-		filter.OrAttribute(predicate);
-		return filter;
+		return @this.Which(Filter.FromPredicate(filter));
 	}
 
 	/// <summary>
-	///     Add additional filters on a <see cref="FieldInfo" /> which has an attribute.
+	///     Filters the applicable <see cref="FieldInfo" /> on which the expectations should be applied.
 	/// </summary>
-	public class WithAttributeFilterResult : Filter.OnField
+	/// <param name="this">The <see cref="IFieldFilter" />.</param>
+	/// <param name="filter">The filter to apply on the <see cref="FieldInfo" />.</param>
+	/// <param name="name">The name of the filter.</param>
+	public static IFieldFilterResult Which(this IFieldFilter @this,
+		Func<FieldInfo, bool> filter,
+		string name)
 	{
-		internal WithAttributeFilterResult(
-			IFieldFilter typeFilter) : base(typeFilter)
-		{
-		}
-
-		/// <summary>
-		///     Adds another filter <see cref="FieldInfo" />s for an attribute of type <typeparamref name="TAttribute" />.
-		/// </summary>
-		/// <param name="predicate">
-		///     (optional) A predicate to check the attribute values.
-		///     <para />
-		///     If not set (<see langword="null" />), will only check if the attribute is present.
-		/// </param>
-		public WithAttributeFilterResult OrAttribute<TAttribute>(
-			Func<TAttribute, bool>? predicate = null) where TAttribute : Attribute
-		{
-			Predicates.Add(Filter.FromPredicate<FieldInfo>(
-				type => type.HasAttribute(predicate),
-				$"Field should have attribute {typeof(TAttribute).Name}"));
-			return this;
-		}
+		return @this.Which(Filter.FromPredicate(filter, name));
 	}
 }
