+++ Source/Testably.Architecture.Rules/Filters/MethodFilterExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq.Expressions;
 using System.Reflection;
 
 namespace Testably.Architecture.Rules;
@@ -6,65 +7,29 @@ namespace Testably.Architecture.Rules;
 /// <summary>
 ///     Extension methods for <see cref="IMethodFilter" />.
 /// </summary>
-public static class MethodFilterExtensions
+public static partial class MethodFilterExtensions
 {
 	/// <summary>
-	///     Filter <see cref="MethodInfo" />s that have an attribute of type <typeparamref name="TAttribute" />.
+	///     Filters the applicable <see cref="MethodInfo" /> on which the expectations should be applied.
 	/// </summary>
 	/// <param name="this">The <see cref="IMethodFilter" />.</param>
-	/// <param name="predicate">
-	///     (optional) A predicate to check the attribute values.
-	///     <para />
-	///     If not set (<see langword="null" />), will only check if the attribute is present.
-	/// </param>
-	/// <param name="inherit">
-	///     <see langword="true" /> to search the inheritance chain to find the attributes; otherwise,
-	///     <see langword="false" />.<br />
-	///     Defaults to <see langword="true" />
-	/// </param>
-	public static WithAttributeFilterResult WithAttribute<TAttribute>(
-		this IMethodFilter @this,
-		Func<TAttribute, bool>? predicate = null,
-		bool inherit = true)
-		where TAttribute : Attribute
+	/// <param name="filter">The filter to apply on the <see cref="MethodInfo" />.</param>
+	public static IMethodFilterResult Which(this IMethodFilter @this,
+		Expression<Func<MethodInfo, bool>> filter)
 	{
-		WithAttributeFilterResult filter = new(
-			@this);
-		filter.OrAttribute(predicate, inherit);
-		return filter;
+		return @this.Which(Filter.FromPredicate(filter));
 	}
 
 	/// <summary>
-	///     Add additional filters on a <see cref="MethodInfo" /> which has an attribute.
+	///     Filters the applicable <see cref="MethodInfo" /> on which the expectations should be applied.
 	/// </summary>
-	public class WithAttributeFilterResult : Filter.OnMethod
+	/// <param name="this">The <see cref="IMethodFilter" />.</param>
+	/// <param name="filter">The filter to apply on the <see cref="MethodInfo" />.</param>
+	/// <param name="name">The name of the filter.</param>
+	public static IMethodFilterResult Which(this IMethodFilter @this,
+		Func<MethodInfo, bool> filter,
+		string name)
 	{
-		internal WithAttributeFilterResult(
-			IMethodFilter typeFilter) : base(typeFilter)
-		{
-		}
-
-		/// <summary>
-		///     Adds another filter <see cref="MethodInfo" />s for an attribute of type <typeparamref name="TAttribute" />.
-		/// </summary>
-		/// <param name="predicate">
-		///     (optional) A predicate to check the attribute values.
-		///     <para />
-		///     If not set (<see langword="null" />), will only check if the attribute is present.
-		/// </param>
-		/// <param name="inherit">
-		///     <see langword="true" /> to search the inheritance chain to find the attributes; otherwise,
-		///     <see langword="false" />.<br />
-		///     Defaults to <see langword="true" />
-		/// </param>
-		public WithAttributeFilterResult OrAttribute<TAttribute>(
-			Func<TAttribute, bool>? predicate = null,
-			bool inherit = true) where TAttribute : Attribute
-		{
-			Predicates.Add(Filter.FromPredicate<MethodInfo>(
-				type => type.HasAttribute(predicate, inherit),
-				$"Method should have attribute {typeof(TAttribute).Name}"));
-			return this;
-		}
+		return @this.Which(Filter.FromPredicate(filter, name));
 	}
 }
